Allow to store directly into a temporary file
On Android we now allow to on-demand store into a temporary file.
There's a chance for a race condition, which I'm not going to take
into account, which is the case where multiple images are been taken
and the file is not been read on time. But it's a tradeoff we need to
accept